refactor: refactor http-interceptor from DI style to function style
