+++ ui/src/app/core/services/interceptors/index.ts
@@ -1,9 +1,3 @@
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { httpInterceptorService } from '@core/services/interceptors/http-interceptor';
 
-import { HttpInterceptorService } from './http-interceptor.service';
-import { LoginExpiredService } from './login-expired.service';
-
-export default [
-  { provide: HTTP_INTERCEPTORS, useClass: HttpInterceptorService, multi: true },
-  { provide: HTTP_INTERCEPTORS, useClass: LoginExpiredService, multi: true }
-];
+export default [httpInterceptorService];
